chore(build): Refactor gruntfile into individual task files.
